Mv RoleMetadata to models.models
