+++ tests/ansible_galaxy/models/test_role_metadata_model.py
@@ -1,6 +1,6 @@
 import logging
 
-from ansible_galaxy.models import role_metadata
+from ansible_galaxy.models.models import RoleMetadata
 
 import attr
 
@@ -8,23 +8,23 @@ log = logging.getLogger(__name__)
 
 
 def test_init_empty():
-    role_md = role_metadata.RoleMetadata()
+    role_md = RoleMetadata()
 
     log.debug('role_md: %s', role_md)
 
-    assert isinstance(role_md, role_metadata.RoleMetadata)
+    assert isinstance(role_md, RoleMetadata)
 
 
 def test_basic():
-    role_md = role_metadata.RoleMetadata(name='some_role',
-                                         author='alikins@redhat.com',
-                                         description='some role that does stuff',
-                                         company='Red Hat',
-                                         license='GPLv3',
-                                         galaxy_tags=['stuff', 'nginx', 'system', 'devel'])
+    role_md = RoleMetadata(name='some_role',
+                           author='alikins@redhat.com',
+                           description='some role that does stuff',
+                           company='Red Hat',
+                           license='GPLv3',
+                           galaxy_tags=['stuff', 'nginx', 'system', 'devel'])
 
     log.debug('role_md: %s', role_md)
-    assert isinstance(role_md, role_metadata.RoleMetadata)
+    assert isinstance(role_md, RoleMetadata)
 
     assert role_md.name == 'some_role'
     assert role_md.author == 'alikins@redhat.com'
@@ -34,9 +34,9 @@ def test_basic():
 
 
 def test_equal():
-    role_md1 = role_metadata.RoleMetadata(name='some_role')
-    role_md1a = role_metadata.RoleMetadata(name='some_role')
-    role_md2 = role_metadata.RoleMetadata(name='a_different_role')
+    role_md1 = RoleMetadata(name='some_role')
+    role_md1a = RoleMetadata(name='some_role')
+    role_md2 = RoleMetadata(name='a_different_role')
 
     assert role_md1 == role_md1a
     assert role_md1a == role_md1
@@ -49,12 +49,12 @@ def test_equal():
 
 
 def test_asdict():
-    role_md = role_metadata.RoleMetadata(name='some_role',
-                                         author='alikins@redhat.com',
-                                         description='some role that does stuff',
-                                         company='Red Hat',
-                                         license='GPLv3',
-                                         galaxy_tags=['stuff', 'nginx', 'system', 'devel'])
+    role_md = RoleMetadata(name='some_role',
+                           author='alikins@redhat.com',
+                           description='some role that does stuff',
+                           company='Red Hat',
+                           license='GPLv3',
+                           galaxy_tags=['stuff', 'nginx', 'system', 'devel'])
 
     log.debug('role_md: %s', role_md)
     role_dict = attr.asdict(role_md)
